Include Jansson and link it
This is based off of reading the Zig Standard Library Documentation and
a build script I saw on the Zig Programming Language Discord. Jansson
includes libc headers, so I set the executable to link libc.

linkSystemLibrary is what was used in the Discord build script, plus
linkLibrary takes a *std.Build.Step.Compile argument which I don't know
how I would get in the first place. I don't know exactly when a linked
library *wouldn't* be needed, but without Jansson this program won't
work so... set it as needed.