Check for leaks with GeneralPurposeAllocator
I wasn't super confident that using the same allocator for
std.json.Reader and std.json.parseFromTokenSource was ok, so I wanted to
check with a GPA. I figure I'll leave it in for now just in case I
accidentally introduce memory leaks.

I'm not sure what the functionName(){} syntax is or why it's needed, but
it was used on [ziglearn.org][1] and without it I get compile errors
unless gpa is const. But if gpa is declared const, then the compiler
complains that gpa.allocator() requires an argument (self: *Self). I
didn't bother trying gpa.allocator(gpa) because I found the ziglearn.org
suggestion, but I imagine that would work too.

[1]: https://ziglearn.org/chapter-2/#allocators