Output parsed design data to a file
Just to make sure that everything I expect to read is there and in the
right format. Given my previous checks on float precision I'm assuming
that the floats are fine even though they're outputted in scientific
notation. But I'll look at the float precision in more detail once I
have to write to a CAD format which I assume requires decimal format.